Use `f32` everywhere (#459)
Fidget's evaluators have always used `f32` values. However, many other
parts of the pipeline (constants in a `Context`, parameters in a shape,
transform matrices, etc) used `f64`s. This was particularly noticeable
in evaluation: the context's tree-walk evaluator used `f64` values, as
did the evaluator test suite. Tiny discrepancies necessitated a bunch of
extra logic (epsilons everywhere, etc).

This PR switches to using `f32` everywhere in Fidget (for things that
eventually end up being evaluated). It then deletes and simplifies a
bunch of awkward hacks and workarounds.

Notably, the evaluator test suite now requires are in *exact agreement*
with the canonical implementation. Unfortunately, the assembly
implementations of `modulo` fell by the wayside; they now call into the
baseline implementation, at slightly higher overhead.

+++ fidget-rhai/src/tree.rs
@@ -14,7 +14,7 @@ impl FromDynamic for Tree {
     ) -> Result<Self, Box<EvalAltResult>> {
         if let Some(t) = d.clone().try_cast::<Tree>() {
             Ok(t)
-        } else if let Ok(v) = f64::from_dynamic(ctx, d.clone(), None) {
+        } else if let Ok(v) = f32::from_dynamic(ctx, d.clone(), None) {
             Ok(Tree::constant(v))
         } else if let Ok(v) = <Vec<Tree>>::from_dynamic(ctx, d.clone(), None) {
             Ok(fidget_shapes::Union { input: v }.into())
